Windows port: normalization: fix narrowing conversions in braced initializers
- groupnorm: explicit static_cast<size_t>/static_cast<uint32_t> for
  int64_t/uint32_t values narrowed into CoreCoord and ttnn::Shape
  braced initializers (groupnorm_grid_utils.cpp, groupnorm_input_mask.cpp)
- layernorm: explicit static_cast<uint32_t> for CoreCoord (size_t) x/y
  fields narrowed into a std::vector<uint32_t> braced return
  (sharded_layernorm_factory_helpers.cpp)
- softmax: explicit static_cast<uint32_t> for size_t block_w/block_h
  products narrowed into DataflowBufferSpec::num_entries
  (softmax_program_factory_attention_optimized_sharded.cpp)

+++ ttnn/cpp/ttnn/operations/normalization/groupnorm/groupnorm_grid_utils.cpp
@@ -130,7 +130,7 @@ GroupNormShardedConfigAndGridSize determine_expected_group_norm_sharded_config_a
         shard_shape = {gn_in_channels_per_core, gn_nhw_per_core};
     }
 
-    const CoreCoord grid_end{static_cast<int>(grid_x) - 1, static_cast<int>(grid_y) - 1};
+    const CoreCoord grid_end{static_cast<size_t>(grid_x) - 1, static_cast<size_t>(grid_y) - 1};
     const CoreRangeSet shard_grid(CoreRange(CoreCoord{0, 0}, grid_end));
 
     const TensorMemoryLayout tensor_memory_layout =
